Fix 'bk commit -Yfile' to not delete file when it is done and to not silently commit with no comments when the file is missing. Add tests to prove this.
bk: 3c5f1ad9z0hjCwDuZ60sWemvEJOwHg

+++ src/commit.c
@@ -55,8 +55,10 @@ commit_main(int ac, char **av)
 				}
 				break;
 		    case 'Y':	doit = 1; getcomment = 0;	/* doc 2.0 */
-				unlink(commentFile);
-				strcpy(commentFile, optarg);
+				if (fileCopy(optarg, commentFile)) {
+					unlink(commentFile);
+					return (1);
+				}
 				break;
 		    case 'A':	/* internal option for regression test only */
 				/* do not document */		/* undoc 2.0 */
